sécurisation des formulaires

+++ lib/pages/UpdateUser.dart
@@ -1,4 +1,5 @@
 import 'dart:developer';
+import 'package:email_validator/email_validator.dart';
 import 'package:flutter/material.dart';
 import 'package:scanqrcode/model/ApiError.dart';
 import 'package:scanqrcode/model/ApiResponse.dart';
@@ -18,6 +19,7 @@ class _UpdateUser extends State<UpdateUser> {
 
   final _formKey = GlobalKey<FormState>();
   ApiResponse _apiResponse;
+  String info = "";
   Key _scaffoldKey;
   final controllerNom = TextEditingController();
   final controllerPrenom = TextEditingController();
@@ -54,6 +56,16 @@ class _UpdateUser extends State<UpdateUser> {
                   Column(
                     crossAxisAlignment: CrossAxisAlignment.start,
                     children: <Widget>[
+                      Center(
+                          child:
+                          Text(
+                            '$info',
+                            style: TextStyle(
+                              fontSize: 18.0,
+                              color: Colors.red,
+                            ),
+                          )
+                      ),
                       TextFormField(
                         key: Key("_nom"),
                         decoration: InputDecoration(labelText: "Nom"),
@@ -109,6 +121,10 @@ class _UpdateUser extends State<UpdateUser> {
                         validator: (value) {
                           if (value.isEmpty) {
                             return 'Mail manquant';
+                          } else {
+                            if (!EmailValidator.validate(value)) {
+                              return "Mail invalide";
+                            }
                           }
                           return null;
                         },
@@ -150,22 +166,29 @@ class _UpdateUser extends State<UpdateUser> {
 
   void _updateUser() async {
     SharedPreferences prefs = await SharedPreferences.getInstance();
-    final FormState form = _formKey.currentState;
-    form.save();
-    _apiResponse = await updateUser(User(nom, prenom, adresse, mail, mdp));
-    if ((_apiResponse.ApiError as ApiError) == null) {
-      if (_apiResponse.UpdateMdp){
-        prefs.clear();
-        Navigator.pushNamedAndRemoveUntil(context, '/login', ModalRoute.withName('/login'));
-      } else {
-        User user =  _apiResponse.Data;
-        prefs.setString("nom", user.nom);
-        prefs.setString("prenom", user.prenom);
-        prefs.setString("mail", user.mail);
-        prefs.setString("adresse", user.adresse);
-        prefs.setString("mdp", user.mdp);
-        Navigator.push(context, MaterialPageRoute(builder: (context) => AppHome()));
+    _formKey.currentState.save();
+    if (_formKey.currentState.validate()) {
+      _apiResponse = await updateUser(User(nom, prenom, adresse, mail, mdp));
+      if ((_apiResponse.ApiError as ApiError) == null) {
+        if (_apiResponse.UpdateMdp) {
+          prefs.clear();
+          Navigator.pushNamedAndRemoveUntil(
+              context, '/login', ModalRoute.withName('/login'));
+        } else {
+          User user = _apiResponse.Data;
+          prefs.setString("nom", user.nom);
+          prefs.setString("prenom", user.prenom);
+          prefs.setString("mail", user.mail);
+          prefs.setString("adresse", user.adresse);
+          prefs.setString("mdp", user.mdp);
+          Navigator.push(
+              context, MaterialPageRoute(builder: (context) => AppHome()));
+        }
       }
+    } else {
+      setState(() {
+        info = "Formulaire invalide";
+      });
     }
   }
 }
